Modified & reviewed  reservation controller

+++ PropertyRentingApi/Controllers/ReservationController.cs
@@ -1,7 +1,8 @@
-﻿using Microsoft.AspNetCore.Http;
+﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using PR_BusinessLayer;
-using PR_DataAccessLayer;
+using SharedDTOLayer.Books.BooksDTO;
+using System.Security.Claims;
 
 namespace PropertyRentingApi.Controllers
 {
@@ -22,7 +23,7 @@ namespace PropertyRentingApi.Controllers
                 return BadRequest($"Not accepted ID {BookId}");
             }
 
-           
+          
 
             PR_BusinessLayer.clsBooking Book = PR_BusinessLayer.clsBooking.Find(BookId);
 
@@ -31,7 +32,7 @@ namespace PropertyRentingApi.Controllers
                 return NotFound($"Student with ID {BookId} not found.");
             }
 
-           
+            
             BookingDTO PDTO = Book.BDTO;
 
             return Ok(PDTO);
@@ -41,20 +42,22 @@ namespace PropertyRentingApi.Controllers
 
 
 
-
+        [Authorize]
         [HttpPost("Add", Name = "AddBook")]
         [ProducesResponseType(StatusCodes.Status201Created)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
-        public ActionResult<BookingDTO> AddBook(BookingDTO ReservationDTO)
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
+        public ActionResult<BookingUpsertDTO> AddBook([FromBody]BookingUpsertDTO ReservationDTO)
         {
+            // note leave (price) field as it is !
             
             if (ReservationDTO == null ||ReservationDTO.PropertyId ==-1 || ReservationDTO.BookedByClientId ==-1 ||
                 ReservationDTO.ExpiredDate == default(DateTime) || ReservationDTO.BeginDate == default(DateTime)
                 
-
                 )
             {
-                return BadRequest("Invalid student data.");
+                return BadRequest("Invalid Reservation data.");
             }
 
             if (ReservationDTO.BeginDate < DateTime.Now || ReservationDTO.ExpiredDate < DateTime.Now) { 
@@ -63,9 +66,9 @@ namespace PropertyRentingApi.Controllers
 
             PR_BusinessLayer.clsProperty PropertyPrice = PR_BusinessLayer.clsProperty.Find(ReservationDTO.PropertyId);
 
-            PR_BusinessLayer.clsBooking reserve = new PR_BusinessLayer.clsBooking(new BookingDTO(ReservationDTO.BookID,ReservationDTO.PropertyId
-                ,ReservationDTO.BookedByClientId ,PropertyPrice.Price , ReservationDTO.ExpiredDate , ReservationDTO.BeginDate, false));
-
+            PR_BusinessLayer.clsBooking reserve = new PR_BusinessLayer.clsBooking(new BookingUpsertDTO(ReservationDTO.BookID,ReservationDTO.PropertyId
+                ,ReservationDTO.BookedByClientId ,PropertyPrice.Price , ReservationDTO.ExpiredDate , ReservationDTO.BeginDate));
+            
             if(ReservationDTO.PropertyId == 0 || ReservationDTO.BookedByClientId == 0)
             {
                 return BadRequest("Invalid PropertyID/ClientId");
@@ -73,9 +76,10 @@ namespace PropertyRentingApi.Controllers
             
            
            PR_BusinessLayer.clsDiscount DiscountProp= PR_BusinessLayer.clsDiscount.GetDiscountsByPropertyID(ReservationDTO.PropertyId);
-            decimal PriceAfterOffer  = clsDiscount.ImplementDiscount(ReservationDTO.PropertyId); 
+            
+            
             if (DiscountProp != null  && DiscountProp.EndDate > DateTime.Now) {
-
+                decimal PriceAfterOffer = clsDiscount.ImplementDiscount(ReservationDTO.PropertyId);
                 reserve.Price= PriceAfterOffer;
              
             }
@@ -86,17 +90,18 @@ namespace PropertyRentingApi.Controllers
             ReservationDTO.Price = reserve.Price;
           
             
-           
+            
+            
             return CreatedAtRoute("AddBook", new { id = ReservationDTO.BookID }, ReservationDTO);
         }
 
 
-        [HttpGet("getActiveReservationDate", Name = "getActiveReservationDate")] // Marks this method to respond to HTTP GET requests.
+        [HttpGet("getActiveReservationDate", Name = "getActiveReservationDate")] 
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
 
-        public ActionResult<ReservationDTO> CheckActiveReservationDate(int PropertyId , DateTime stDate , DateTime exDate)
+        public ActionResult<CheckForActiveReservationDTO> CheckActiveReservationDate(int PropertyId , DateTime stDate , DateTime exDate)
         {
 
             if (PropertyId < 0)
@@ -111,27 +116,31 @@ namespace PropertyRentingApi.Controllers
                 return NotFound($"Book with ID {PropertyId} not found or No Valid Offer is exist for this Property.");
             }
 
-
-
             return Ok(ActiveReservationDate);
         }
-
+        [Authorize]
         [HttpPut("UpdateReservationDate", Name = "UpdateReservationDate")]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
-        public ActionResult<BookingUpdateDTO> UpdateStudent(int UpdateReservationById, BookingUpdateDTO ReservationDTO)
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
+        public ActionResult<BookingUpsertDTO> UpdateReservationDate(BookingUpdateDTO ReservationDTO)
         {
             if (ReservationDTO == null || ReservationDTO.ExpiredDate == default(DateTime) || ReservationDTO.BeginDate == default(DateTime))
             {
                 return BadRequest("Invalid student data.");
             }
 
+            var userIdStr = User.FindFirst(ClaimTypes.NameIdentifier)!.Value;
+            int clientReservation = PR_BusinessLayer.clsBooking.Find(ReservationDTO.BookID).BookedByClientId;
+
+            if (Convert.ToInt32(userIdStr) != clientReservation) return Unauthorized("you're Unauthorized to update this Reservation");
 
 
 
 
-            PR_BusinessLayer.clsBooking reserve = PR_BusinessLayer.clsBooking.Find(UpdateReservationById);
+            PR_BusinessLayer.clsBooking reserve = PR_BusinessLayer.clsBooking.Find(ReservationDTO.BookID);
             if (reserve == null )
             {
                 return BadRequest("Invalid Reservation ID");
@@ -145,16 +154,18 @@ namespace PropertyRentingApi.Controllers
 
             reserve.Save();
 
-            //we return the DTO not the full student object.
+            
             return Ok(reserve.BUDTO);
 
         }
 
-
+        [Authorize]
         [HttpDelete("DeleteReservation", Name = "DeleteReservation")]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
         public ActionResult DeleteReservationType(int BookId)
         {
             if (BookId < 1)
@@ -162,7 +173,12 @@ namespace PropertyRentingApi.Controllers
                 return BadRequest($"Not accepted ID {BookId}");
             }
 
-            
+            var userIdStr = User.FindFirst(ClaimTypes.NameIdentifier)!.Value;
+            int clientReservation = PR_BusinessLayer.clsBooking.Find(BookId).BookedByClientId;
+
+            if (Convert.ToInt32(userIdStr) != clientReservation) return Unauthorized("you're Unauthorized to delete this Reservation");
+
+
 
             if (PR_BusinessLayer.clsBooking.DeleteBooking(BookId) != -1)
